Fixed Inventory::getItem always returning null
Simple fix, changed the auto expression in getItem's loop to an auto reference
this stopped objects being created in the loop, which were being destroyed
at the end of the loop and thus caused the null pointer check block in
Unit::useItem to be executed.

Additionally, Inventory::removeItem was simplified to a lambda expression.